+++ examples/data-sources/morpheus_contact/data-source.tf
@@ -1,0 +1,3 @@
+data "morpheus_contact" "example" {
+  name = "example contact"
+}