+++ examples/resources/morpheus_contact/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_contact.tf_example_contact 1