+++ CHANGELOG.md
@@ -13,9 +13,11 @@ NOTES:
 
 FEATURES:
 
+* **New Data Source:** `morpheus_contact`
 * **New Data Source:** `morpheus_execute_schedule`
 * **New Data Source:** `morpheus_tenant_role`
 * **New Data Source:** `morpheus_workflow`
+* **New Resource:** `morpheus_contact`
 * **New Resource:** `morpheus_execute_schedule`
 * **New Resource:** `morpheus_max_cores_policy`
 * **New Resource:** `morpheus_max_hosts_policy`